+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## Change Log
 All notable changes to this project will be documented in this file.
 
+### [[Unreleased](https://github.com/Zeroji/semicolon/releases/latest)]
+
+#### Fixed
+
+* `text || ;command` syntax didn't work
+
 ### [[v0.2.0](https://github.com/Zeroji/semicolon/releases/tag/v0.2.0)]
 
 #### Added